feat: Add feature flags for reference and optimized Pauli gate implementations
This commit introduces the `optimized` feature flag in Cargo.toml and adds conditional
compilation for X, Y, and Z Pauli gates.

This allows the reference implementation to be tested while also
optimising for the Pauli gates which are based on whole multiples of Ix,
Iy or Iz so end up just switching coefficients, just like the standard
Pauli matrices.

The optimized version can be run against the same unit tests with:

cargo test --features optimized